Split rasterizer into multiple files.
